Improve activity thread hooking
