+++ booster-android-instrument-thread/src/main/java/com/didiglobal/booster/instrument/ShadowAsyncTask.java
@@ -4,7 +4,7 @@ import android.util.Log;
 
 import java.util.concurrent.ThreadPoolExecutor;
 
-import static com.didiglobal.booster.android.bugfix.Constants.TAG;
+import static com.didiglobal.booster.instrument.Constants.TAG;
 
 /**
  * @author neighbWang
