+++ booster-android-instrument/src/main/java/com/didiglobal/booster/instrument/CaughtCallback.java
@@ -1,4 +1,4 @@
-package com.didiglobal.booster.android.bugfix;
+package com.didiglobal.booster.instrument;
 
 import android.os.Handler;
 import android.os.Message;
