+++ booster-android-instrument/src/main/java/com/didiglobal/booster/instrument/Constants.java
@@ -1,4 +1,4 @@
-package com.didiglobal.booster.android.bugfix;
+package com.didiglobal.booster.instrument;
 
 public interface Constants {
