+++ booster-android-instrument/src/main/java/com/didiglobal/booster/instrument/Reflection.java
@@ -1,4 +1,4 @@
-package com.didiglobal.booster.android.bugfix;
+package com.didiglobal.booster.instrument;
 
 import android.util.Log;
